Removing old queries and results
