+++ src/util/DataStores/Graph.rs
@@ -2,11 +2,15 @@
 use super::{
   TripleStore::{TripleStore, TripleStoreRefIterator},
   super::{
-    Ordering, Triple, Double, QueryDouble, QueryTriple,
-    Queries::{Query, QueryUnit},
-    Results::{Result, ResultUnit, ResultCollection}
+    Ordering
+    // Results::{Result, ResultUnit, ResultCollection}
   }
 };
+
+type Triple = (String, String, String);
+type QueryTriple = (Option<String>, Option<String>, Option<String>);
+// type Double = (String, String);
+// type QueryDouble = (Option<String>, Option<String>);
 
 /*
 A data-structure that sacrifices space for fast data access
@@ -128,201 +132,201 @@ impl Graph {
       },
     }
   }
-  fn get_double(&self, qd: &QueryDouble, ord: [Ordering; 2]) -> Vec<Double> {
-    use Ordering::{S, P, O};
-    let store = match &ord {
-      [S,P] => { &self.spo },
-      [P,O] => { &self.pos },
-      [O,S] => { &self.osp },
-      _  => { return Vec::new() },
-    };
-    store.get_double(qd)
-  }
-  fn get_single(&self, q: &Option<String>, ord: Ordering) -> Vec<String> {
-    use Ordering::{S, P, O};
-    let store = match &ord {
-      S => { &self.spo },
-      P => { &self.pos },
-      O => { &self.osp },
-      _  => { return Vec::new() },
-    };
-    store.get_single(q)
-  }
-  pub fn get(&self, q: Query) -> ResultCollection {
-    use QueryUnit::{Val, Var, Anon, Ignore};
-    use Ordering::{S, P, O};
-    let mut rc = ResultCollection::new();
-    rc.query = q.clone();
-    match q {
-      Query::Null => {/*Do nothing*/},
-      Query::Single(h, ord) => {
-        //Filter out all the Ignores, call again with corrected query
-        match &h {
-          Ignore => {
-            return rc
-          },
-          _ => {},
-        };
-        //Actually start processing now
-        let mut q: Option<String>;
-        match &h {
-          Val(a) => { q = Some(a.clone()); },
-          Var(_)
-          | Anon
-          | Ignore => { q = None; },
-        };
-        //Rearrange the Ordering to match the stores Graph has
-        let query_res = match &ord {
-          P => { self.get_single(&q, P) },
-          S => { self.get_single(&q, S) },
-          O => { self.get_single(&q, O) },
-          _ => { self.get_single(&q, S) },
-        };
-        if query_res.len() > 0 {
-          for i in 0..query_res.len() {
-            let mut r = Result::new();
-            match &h {
-              Val(a) => { r.add_anon(ResultUnit::Value(a.to_string())); },
-              Var(a) => { r.add_var(a.to_string(), query_res[i].clone()); },
-              Anon   => { r.add_anon(ResultUnit::Value(query_res[i].clone())); },
-              Ignore => { r.add_anon(ResultUnit::Ignore); },
-            }
-            rc.results.push(r);
-          }
-        }
-      },
-      Query::Double(h, t, ord) => {
-        //Filter out all the Ignores, call again with corrected query
-        match (&h, &t) {
-          (Ignore, Ignore) => {
-            return rc
-          },
-          (Ignore, _) => {
-            return self.get(Query::Single(t, ord[1].clone()))
-          },
-          (_, Ignore) => {
-            return self.get(Query::Single(h, ord[0].clone()))
-          },
-          _ => {},
-        };
-        //Actually start processing now
-        let mut q1: Option<String>;
-        let mut q2: Option<String>;
-        match &h {
-          Val(a) => { q1 = Some(a.clone()); },
-          Var(_)
-          | Anon
-          | Ignore => { q1 = None; },
-        };
-        match &t {
-          Val(b) => { q2 = Some(b.clone()); },
-          Var(_)
-          | Anon
-          | Ignore => { q2 = None; },
-        };
-        //Rearrange the Ordering to match the stores Graph has
-        let query_res = match &ord {
-          [P,S] => { self.get_double(&(q1, q2), [S,P]) },
-          [S,O] => { self.get_double(&(q1, q2), [O,S]) },
-          [O,P] => { self.get_double(&(q1, q2), [P,O]) },
-          _ => { self.get_double(&(q1, q2), ord) }
-        };
-        if query_res.len() > 0 {
-          for i in 0..query_res.len() {
-            let mut r = Result::new();
-            match &h {
-              Val(a) => { r.add_anon(ResultUnit::Value(a.to_string())); },
-              Var(a) => { r.add_var(a.to_string(), query_res[i].0.clone()); },
-              Anon   => { r.add_anon(ResultUnit::Value(query_res[i].0.clone())); },
-              Ignore => { r.add_anon(ResultUnit::Ignore); },
-            }
-            match &t {
-              Val(b) => { r.add_anon(ResultUnit::Value(b.to_string())); },
-              Var(b) => { r.add_var(b.to_string(), query_res[i].1.clone()); },
-              Anon   => { r.add_anon(ResultUnit::Value(query_res[i].1.clone())); },
-              Ignore => { r.add_anon(ResultUnit::Ignore); },
-            }
-            rc.results.push(r);
-          }
-        }
-      },
-      Query::Triple(s, p, o, _) => {
-        //Filter out all the Ignores, call again with corrected query
-        match (&s, &p, &o) {
-          (Ignore, Ignore, Ignore) => {
-            return rc
-          },
-          (Ignore, Ignore, _) => {
-            return self.get(Query::Single(o, O))
-          },
-          (Ignore, _, Ignore) => {
-            return self.get(Query::Single(p, P))
-          },
-          (_, Ignore, Ignore) => {
-            return self.get(Query::Single(s, S))
-          },
-          (Ignore, _, _) => {
-            return self.get(Query::Double(p, o, [P, O]))
-          },
-          (_, Ignore, _) => {
-            return self.get(Query::Double(s, o, [S, O]))
-          },
-          (_, _, Ignore) => {
-            return self.get(Query::Double(s, p, [S, P]))
-          },
-          _ => {},
-        };
-        //Actually start processing now
-        let mut q1: Option<String>;
-        let mut q2: Option<String>;
-        let mut q3: Option<String>;
-        match &s {
-          Val(a) => { q1 = Some(a.clone()); },
-          Var(_)
-          | Anon
-          | Ignore => { q1 = None; },
-        };
-        match &p {
-          Val(b) => { q2 = Some(b.clone()); },
-          Var(_)
-          | Anon
-          | Ignore => { q2 = None; },
-        };
-        match &o {
-          Val(b) => { q3 = Some(b.clone()); },
-          Var(_)
-          | Anon
-          | Ignore => { q3 = None; },
-        };
-        let query_res = self.get_triple(&(q1, q2, q3));
-        if query_res.len() > 0 {
-          for i in 0..query_res.len() {
-            let mut r = Result::new();
-            match &s {
-              Val(a) => { r.add_anon(ResultUnit::Value(a.to_string())); },
-              Var(a) => { r.add_var(a.to_string(), query_res[i].0.clone()); },
-              Anon   => { r.add_anon(ResultUnit::Value(query_res[i].0.clone())); },
-              Ignore => { r.add_anon(ResultUnit::Ignore); },
-            }
-            match &p {
-              Val(b) => { r.add_anon(ResultUnit::Value(b.to_string())); },
-              Var(b) => { r.add_var(b.to_string(), query_res[i].1.clone()); },
-              Anon   => { r.add_anon(ResultUnit::Value(query_res[i].1.clone())); },
-              Ignore => { r.add_anon(ResultUnit::Ignore); },
-            }
-            match &o {
-              Val(c) => { r.add_anon(ResultUnit::Value(c.to_string())); },
-              Var(c) => { r.add_var(c.to_string(), query_res[i].2.clone()); },
-              Anon   => { r.add_anon(ResultUnit::Value(query_res[i].2.clone())); },
-              Ignore => { r.add_anon(ResultUnit::Ignore); },
-            }
-            rc.results.push(r);
-          }
-        }
-      },
-    };
-    rc
-  }
+  // fn get_double(&self, qd: &QueryDouble, ord: [Ordering; 2]) -> Vec<Double> {
+  //   use Ordering::{S, P, O};
+  //   let store = match &ord {
+  //     [S,P] => { &self.spo },
+  //     [P,O] => { &self.pos },
+  //     [O,S] => { &self.osp },
+  //     _  => { return Vec::new() },
+  //   };
+  //   store.get_double(qd)
+  // }
+  // fn get_single(&self, q: &Option<String>, ord: Ordering) -> Vec<String> {
+  //   use Ordering::{S, P, O};
+  //   let store = match &ord {
+  //     S => { &self.spo },
+  //     P => { &self.pos },
+  //     O => { &self.osp },
+  //     _  => { return Vec::new() },
+  //   };
+  //   store.get_single(q)
+  // }
+  // pub fn get(&self, q: Query) -> ResultCollection {
+  //   use QueryUnit::{Val, Var, Anon, Ignore};
+  //   use Ordering::{S, P, O};
+  //   let mut rc = ResultCollection::new();
+  //   rc.query = q.clone();
+  //   match q {
+  //     Query::Null => {/*Do nothing*/},
+  //     Query::Single(h, ord) => {
+  //       //Filter out all the Ignores, call again with corrected query
+  //       match &h {
+  //         Ignore => {
+  //           return rc
+  //         },
+  //         _ => {},
+  //       };
+  //       //Actually start processing now
+  //       let mut q: Option<String>;
+  //       match &h {
+  //         Val(a) => { q = Some(a.clone()); },
+  //         Var(_)
+  //         | Anon
+  //         | Ignore => { q = None; },
+  //       };
+  //       //Rearrange the Ordering to match the stores Graph has
+  //       let query_res = match &ord {
+  //         P => { self.get_single(&q, P) },
+  //         S => { self.get_single(&q, S) },
+  //         O => { self.get_single(&q, O) },
+  //         _ => { self.get_single(&q, S) },
+  //       };
+  //       if query_res.len() > 0 {
+  //         for i in 0..query_res.len() {
+  //           let mut r = Result::new();
+  //           match &h {
+  //             Val(a) => { r.add_anon(ResultUnit::Value(a.to_string())); },
+  //             Var(a) => { r.add_var(a.to_string(), query_res[i].clone()); },
+  //             Anon   => { r.add_anon(ResultUnit::Value(query_res[i].clone())); },
+  //             Ignore => { r.add_anon(ResultUnit::Ignore); },
+  //           }
+  //           rc.results.push(r);
+  //         }
+  //       }
+  //     },
+  //     Query::Double(h, t, ord) => {
+  //       //Filter out all the Ignores, call again with corrected query
+  //       match (&h, &t) {
+  //         (Ignore, Ignore) => {
+  //           return rc
+  //         },
+  //         (Ignore, _) => {
+  //           return self.get(Query::Single(t, ord[1].clone()))
+  //         },
+  //         (_, Ignore) => {
+  //           return self.get(Query::Single(h, ord[0].clone()))
+  //         },
+  //         _ => {},
+  //       };
+  //       //Actually start processing now
+  //       let mut q1: Option<String>;
+  //       let mut q2: Option<String>;
+  //       match &h {
+  //         Val(a) => { q1 = Some(a.clone()); },
+  //         Var(_)
+  //         | Anon
+  //         | Ignore => { q1 = None; },
+  //       };
+  //       match &t {
+  //         Val(b) => { q2 = Some(b.clone()); },
+  //         Var(_)
+  //         | Anon
+  //         | Ignore => { q2 = None; },
+  //       };
+  //       //Rearrange the Ordering to match the stores Graph has
+  //       let query_res = match &ord {
+  //         [P,S] => { self.get_double(&(q1, q2), [S,P]) },
+  //         [S,O] => { self.get_double(&(q1, q2), [O,S]) },
+  //         [O,P] => { self.get_double(&(q1, q2), [P,O]) },
+  //         _ => { self.get_double(&(q1, q2), ord) }
+  //       };
+  //       if query_res.len() > 0 {
+  //         for i in 0..query_res.len() {
+  //           let mut r = Result::new();
+  //           match &h {
+  //             Val(a) => { r.add_anon(ResultUnit::Value(a.to_string())); },
+  //             Var(a) => { r.add_var(a.to_string(), query_res[i].0.clone()); },
+  //             Anon   => { r.add_anon(ResultUnit::Value(query_res[i].0.clone())); },
+  //             Ignore => { r.add_anon(ResultUnit::Ignore); },
+  //           }
+  //           match &t {
+  //             Val(b) => { r.add_anon(ResultUnit::Value(b.to_string())); },
+  //             Var(b) => { r.add_var(b.to_string(), query_res[i].1.clone()); },
+  //             Anon   => { r.add_anon(ResultUnit::Value(query_res[i].1.clone())); },
+  //             Ignore => { r.add_anon(ResultUnit::Ignore); },
+  //           }
+  //           rc.results.push(r);
+  //         }
+  //       }
+  //     },
+  //     Query::Triple(s, p, o, _) => {
+  //       //Filter out all the Ignores, call again with corrected query
+  //       match (&s, &p, &o) {
+  //         (Ignore, Ignore, Ignore) => {
+  //           return rc
+  //         },
+  //         (Ignore, Ignore, _) => {
+  //           return self.get(Query::Single(o, O))
+  //         },
+  //         (Ignore, _, Ignore) => {
+  //           return self.get(Query::Single(p, P))
+  //         },
+  //         (_, Ignore, Ignore) => {
+  //           return self.get(Query::Single(s, S))
+  //         },
+  //         (Ignore, _, _) => {
+  //           return self.get(Query::Double(p, o, [P, O]))
+  //         },
+  //         (_, Ignore, _) => {
+  //           return self.get(Query::Double(s, o, [S, O]))
+  //         },
+  //         (_, _, Ignore) => {
+  //           return self.get(Query::Double(s, p, [S, P]))
+  //         },
+  //         _ => {},
+  //       };
+  //       //Actually start processing now
+  //       let mut q1: Option<String>;
+  //       let mut q2: Option<String>;
+  //       let mut q3: Option<String>;
+  //       match &s {
+  //         Val(a) => { q1 = Some(a.clone()); },
+  //         Var(_)
+  //         | Anon
+  //         | Ignore => { q1 = None; },
+  //       };
+  //       match &p {
+  //         Val(b) => { q2 = Some(b.clone()); },
+  //         Var(_)
+  //         | Anon
+  //         | Ignore => { q2 = None; },
+  //       };
+  //       match &o {
+  //         Val(b) => { q3 = Some(b.clone()); },
+  //         Var(_)
+  //         | Anon
+  //         | Ignore => { q3 = None; },
+  //       };
+  //       let query_res = self.get_triple(&(q1, q2, q3));
+  //       if query_res.len() > 0 {
+  //         for i in 0..query_res.len() {
+  //           let mut r = Result::new();
+  //           match &s {
+  //             Val(a) => { r.add_anon(ResultUnit::Value(a.to_string())); },
+  //             Var(a) => { r.add_var(a.to_string(), query_res[i].0.clone()); },
+  //             Anon   => { r.add_anon(ResultUnit::Value(query_res[i].0.clone())); },
+  //             Ignore => { r.add_anon(ResultUnit::Ignore); },
+  //           }
+  //           match &p {
+  //             Val(b) => { r.add_anon(ResultUnit::Value(b.to_string())); },
+  //             Var(b) => { r.add_var(b.to_string(), query_res[i].1.clone()); },
+  //             Anon   => { r.add_anon(ResultUnit::Value(query_res[i].1.clone())); },
+  //             Ignore => { r.add_anon(ResultUnit::Ignore); },
+  //           }
+  //           match &o {
+  //             Val(c) => { r.add_anon(ResultUnit::Value(c.to_string())); },
+  //             Var(c) => { r.add_var(c.to_string(), query_res[i].2.clone()); },
+  //             Anon   => { r.add_anon(ResultUnit::Value(query_res[i].2.clone())); },
+  //             Ignore => { r.add_anon(ResultUnit::Ignore); },
+  //           }
+  //           rc.results.push(r);
+  //         }
+  //       }
+  //     },
+  //   };
+  //   rc
+  // }
 }
 
 fn t_order(t: Triple, curr_ordering: &Ordering) -> Triple {
