+++ src/util/SPARQL/mod.rs
@@ -1,1 +1,2 @@
 pub mod Query;
+pub mod Result;
